+++ Machines/Atari2600/TIA.hpp
@@ -144,8 +144,8 @@ class TIA {
 		// movement
 		bool horizontal_blank_extend_;
 		int horizontal_move_start_time_;
-		uint8_t motion_[5];
-		uint8_t position_[5];
+		int motion_[5];
+		int position_[5];
 		bool is_moving_[5];
 		enum class MotionIndex : uint8_t {
 			Ball,
